--- src/main/java/net/detalk/api/controller/v1/request/UpdateProductPostRequest.java
@@ -1,16 +1,0 @@
-package net.detalk.api.controller.v1.request;
-
-import java.util.List;
-import lombok.Builder;
-
-@Builder
-public record UpdateProductPostRequest(
-    String name,
-    String pricingPlan,
-    String description,
-    List<String> tags,
-    String url,
-    List<String> imageIds,
-    boolean isMaker
-) {
-}
