--- src/main/java/net/detalk/api/controller/v1/response/CreateProductPostResponse.java
@@ -1,4 +1,0 @@
-package net.detalk.api.controller.v1.response;
-
-public record CreateProductPostResponse(Long id) {
-}
