+++ src/main/java/net/detalk/api/post/controller/response/CreateProductPostResponse.java
@@ -1,0 +1,9 @@
+package net.detalk.api.post.controller.response;
+
+/**
+ * 상품 포스트 생성 응답 DTO
+ *
+ * @param id 생성된 상품 포스트 ID
+ */
+public record CreateProductPostResponse(Long id) {
+}
