+++ src/main/java/net/detalk/api/post/domain/ProductPostSnapshotTag.java
@@ -1,4 +1,4 @@
-package net.detalk.api.domain;
+package net.detalk.api.post.domain;
 
 import lombok.Builder;
 import lombok.Getter;
