+++ src/main/java/net/detalk/api/post/domain/exception/ProductPostSnapshotUpdateException.java
@@ -1,4 +1,4 @@
-package net.detalk.api.domain.exception;
+package net.detalk.api.post.domain.exception;
 
 import net.detalk.api.support.error.ApiException;
 import org.springframework.http.HttpStatus;
