+++ src/main/java/net/detalk/api/post/repository/ProductPostSnapshotAttachmentFileRepository.java
@@ -1,0 +1,9 @@
+package net.detalk.api.post.repository;
+
+import net.detalk.api.post.domain.ProductPostSnapshotAttachmentFile;
+
+public interface ProductPostSnapshotAttachmentFileRepository {
+
+    ProductPostSnapshotAttachmentFile save(ProductPostSnapshotAttachmentFile attachmentFile);
+
+}
