+++ src/main/java/net/detalk/api/post/repository/impl/ProductPostLastSnapshotRepositoryImpl.java
@@ -1,9 +1,10 @@
-package net.detalk.api.repository;
+package net.detalk.api.post.repository.impl;
 
 import java.util.Optional;
 import lombok.RequiredArgsConstructor;
-import net.detalk.api.domain.ProductPostLastSnapshot;
-import net.detalk.api.domain.ProductPostSnapshot;
+import net.detalk.api.post.domain.ProductPostLastSnapshot;
+import net.detalk.api.post.domain.ProductPostSnapshot;
+import net.detalk.api.post.repository.ProductPostLastSnapshotRepository;
 import org.jooq.DSLContext;
 import org.springframework.stereotype.Repository;
 
@@ -11,10 +12,11 @@ import static net.detalk.jooq.tables.JProductPostLastSnapshot.PRODUCT_POST_LAST_
 
 @RequiredArgsConstructor
 @Repository
-public class ProductPostLastSnapshotRepository {
+public class ProductPostLastSnapshotRepositoryImpl implements ProductPostLastSnapshotRepository {
 
     private final DSLContext dsl;
 
+    @Override
     public ProductPostLastSnapshot save(Long postId, Long snapshotId) {
         return dsl.insertInto(PRODUCT_POST_LAST_SNAPSHOT)
             .set(PRODUCT_POST_LAST_SNAPSHOT.POST_ID, postId)
@@ -23,12 +25,14 @@ public class ProductPostLastSnapshotRepository {
             .fetchOneInto(ProductPostLastSnapshot.class);
     }
 
+    @Override
     public Optional<ProductPostLastSnapshot> findByPostId(Long postId) {
         return dsl.selectFrom(PRODUCT_POST_LAST_SNAPSHOT)
             .where(PRODUCT_POST_LAST_SNAPSHOT.POST_ID.eq(postId))
             .fetchOptionalInto(ProductPostLastSnapshot.class);
     }
 
+    @Override
     public int update(Long postId, ProductPostSnapshot newSnapshot) {
         return dsl.update(PRODUCT_POST_LAST_SNAPSHOT)
             .set(PRODUCT_POST_LAST_SNAPSHOT.SNAPSHOT_ID, newSnapshot.getId())
