+++ src/main/java/net/detalk/api/post/repository/impl/ProductPostLinkRepositoryImpl.java
@@ -1,18 +1,20 @@
-package net.detalk.api.repository;
+package net.detalk.api.post.repository.impl;
 
 import static net.detalk.jooq.tables.JProductPostLink.PRODUCT_POST_LINK;
 
 import lombok.RequiredArgsConstructor;
-import net.detalk.api.domain.ProductPostLink;
+import net.detalk.api.post.domain.ProductPostLink;
+import net.detalk.api.post.repository.ProductPostLinkRepository;
 import org.jooq.DSLContext;
 import org.springframework.stereotype.Repository;
 
 @RequiredArgsConstructor
 @Repository
-public class ProductPostLinkRepository {
+public class ProductPostLinkRepositoryImpl implements ProductPostLinkRepository {
 
     private final DSLContext dsl;
 
+    @Override
     public ProductPostLink save(Long postId, Long linkId) {
         return dsl.insertInto(PRODUCT_POST_LINK)
             .set(PRODUCT_POST_LINK.POST_ID, postId)
