+++ src/test/java/net/detalk/api/service/RecommendServiceTest.java
@@ -19,8 +19,9 @@ import java.util.Optional;
 import net.detalk.api.controller.v1.request.CreateRecommendRequest;
 import net.detalk.api.domain.Recommend;
 import net.detalk.api.domain.exception.DuplicateRecommendationException;
-import net.detalk.api.domain.exception.ProductPostNotFoundException;
+import net.detalk.api.post.domain.exception.ProductPostNotFoundException;
 import net.detalk.api.mock.FakeTimeHolder;
+import net.detalk.api.post.service.ProductPostService;
 import net.detalk.api.repository.RecommendProductRepository;
 import net.detalk.api.repository.RecommendRepository;
 import net.detalk.api.support.TimeHolder;
